requirements.txt
numpy<2.0
matplotlib==3.7.1
pyserial==3.5
